VideoBox component로 나누기
