+++ frontend/src/components/VideoBox/ThumbNail/ThumbNail.scss
