+++ frontend/src/components/VideoBox/ThumbNail/index.tsx
