+++ frontend/src/components/VideoBox/VideoBox.scss
@@ -1,0 +1,63 @@
+// TODO scss 변수 사용해서 간결화하기
+.video {
+  display: flex;
+  flex-direction: column;
+  margin: 0 8px 40px 8px;
+  width: 80vw;
+
+  &:hover {
+    .thumb_nail {
+      animation: fadein 2s;
+      animation-fill-mode: forwards;
+
+      @keyframes fadein {
+        from {
+          opacity: 0;
+        }
+        to {
+          opacity: 1;
+        }
+      }
+    }
+
+    .run_time {
+      animation: fadeout 1s;
+      animation-fill-mode: forwards;
+      @keyframes fadeout {
+        from {
+          opacity: 1;
+        }
+        to {
+          opacity: 0;
+        }
+      }
+    }
+
+    .hidden {
+      .lastview {
+        visibility: visible;
+        transition: width 2s;
+        &:hover .lastview-item1 {
+          display: block;
+          .overlay1 {
+            transform: translateX(0);
+          }
+        }
+      }
+
+      .addlist {
+        visibility: visible;
+        &:hover .addlist-item1 {
+          display: block;
+          .overlay2 {
+            transform: translateX(0);
+          }
+        }
+      }
+    }
+
+    .hiddenAdd > .add {
+      visibility: visible;
+    }
+  }
+}
