+++ frontend/src/components/VideoBox/VideoDesc/VideoDesc.scss
