+++ frontend/src/components/VideoContainer/VideoContainer.scss
@@ -18,72 +18,4 @@ p {
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  // width: calc(80% - 16px);
-  // height: 60vw;
-}
-
-// VideoBox element
-.video {
-  display: flex;
-  flex-direction: column;
-  margin: 0 8px 40px 8px;
-  // TODO 다시 80으로 복귀
-  width: 80vw;
-
-  &:hover {
-    .thumb_nail {
-      // background-image: url(./images/에러.png);
-      animation: fadein 2s;
-      animation-fill-mode: forwards;
-
-      @keyframes fadein {
-        from {
-          opacity: 0;
-        }
-        to {
-          opacity: 1;
-        }
-      }
-    }
-
-    .run_time {
-      animation: fadeout 1s;
-      animation-fill-mode: forwards;
-      @keyframes fadeout {
-        from {
-          opacity: 1;
-        }
-        to {
-          opacity: 0;
-        }
-      }
-    }
-
-    .hidden {
-      .lastview {
-        visibility: visible;
-        transition: width 2s;
-        &:hover .lastview-item1 {
-          display: block;
-          .overlay1 {
-            transform: translateX(0);
-          }
-        }
-      }
-
-      .addlist {
-        visibility: visible;
-        &:hover .addlist-item1 {
-          display: block;
-          .overlay2 {
-            transform: translateX(0);
-          }
-        }
-      }
-    }
-
-    .hiddenAdd > .add {
-      visibility: visible;
-    }
-  }
 }
